Remove the MediaManagerToRTPSessionConnector
