+++ examples/Softphone/SIPSorcery.SoftPhone/SIPSorcery.SoftPhone.csproj
@@ -101,7 +101,6 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="AppState.cs" />
-    <Compile Include="Media\MediaManagerToRTPSessionConnector.cs" />
     <Compile Include="Media\RTPMediaSessionFactory.cs" />
     <Compile Include="Media\VideoSubTypes.cs" />
     <Compile Include="Signalling\IVoIPClient.cs" />
